refactor: pass all objects in single method
